Reimplemented cursor and fullscan cursor.
Now ICursor is just interface and all implementation goes to its inheritors.
Also added method HasNext to check if there is more values to get.
