+++ src/core/Cursor.h
@@ -1,31 +1,14 @@
 #ifndef Cursor_h
 #define Cursor_h
 
-#include <memory>
-#include <vector>
-
-#include "DataPage.h"
 #include "Record.h"
 
-struct ColumnDescriptor;
-struct PageManager;
-
 struct ICursor {
-	ICursor(PageManager & pageManager, PageID startPageID, ColumnDescriptors const & descriptors);
 	virtual ~ICursor() = default;
-
-	Record Get();
+	virtual Record Get() = 0;
 	/// Returns true if there is next record and moves to it.
-	virtual bool Next();
-
-protected:
-	virtual bool AcceptCurrentRecord() = 0;
-
-protected:
-	ColumnDescriptors const & m_descriptors;
-	PageManager & m_pageManager;
-	std::unique_ptr<DataPage> m_currentPage;
-	size_t m_currentRecordNumber; /// in page
+	virtual bool Next() = 0;
+	virtual bool HasNext() const = 0;
 };
 
 #endif // Cursor_h
